Work mostly with IOBuffers
Instead of creating IOBuffers within each function, pass around IOBuffers to modifying functions and only take from them at the end of the process!  This vastly reduces the allocations as we are no longer creating a new IOBuffer within each function.  However, it is still not as fast as it was (but its memory usage is better).