Implement spelled_out for numbers up to 999 for Māori
I'm not sure how I was aiming to implement it previously, as it has
been a while since I've worked on this project, but I implemented
numbers up to 999 this evening.  It certainly doesn't feel as though I
am doing this the most efficient way, but premature optimisation is
the root of all evil; I can optimise once it is working.  Overall,
this commit has made good progress.