Fix initial word by accounting for order of magnitude in mi
In Māori's (mi) implementation of spelled_out, there is a special case
in which 1xx and 1xxx numbers start with _ko_tahi (rather than just
tahi, which is the equivalent of one).  If the number is > 99 and is
not 1xx or 1xxx, we need to use the first digit.  I initially wrote
the first digit as d being n ÷ 10, however this is not correct when n
> 1000, so we have to account for order of magnitude