Set diffs to modify previous countmap in self-referencing example
When we calculate the diffs from previous to current countmaps in
self-referencing example, we previously created a new dictionary.  As
we never use the previous countmap after this point, we can simply
modify it!

This seems to speed things up immensely:
 Without mutation of previous:
     julia> @btime construct_true_pangram();
       8.100 s (65276305 allocations: 4.59 GiB)

  With mutation of previous:
     julia> @btime construct_true_pangram();
       2.992 s (26294381 allocations: 1.67 GiB)

Next steps for optimisation would be to reuse the same buffer for
pangram construction.