Remove initial state from static chars in self-referencing example
While it is true that certain characters in this sentence are never
used in spelling out numbers, creating an initial state accounting for
these characters has not proven useful, and hence this commit removes
this logic.

I'm not entirely sure why, but for some reason, having this initial
state considerably slows down the programme.