More progress on spelling out postive integers in Te Reo (Māori)
Requires heavy cleaning, but still just uncovering patterns etc.  In
fact, the basic implementation is nearly complete!  Aside from a stray
comma that may or may not be needed, and some larger number scales,
this is done.  Then I will be able to clean it up and decouple some
patterns.

I'm really pleased with how this is turning out.  It's neat that I
can't find any implementation of this before now, and I'm pleased that
it's almost finished, despite my busy schedule (I haven't worked on it
in a while, but picked it up again tonight in the spirit of
procrastination).

Once the basic implementation is done, I will be able to ensure this
works for negative numbers, and possibly ordinal numbers (we will
see).  After that I should be able to merge it into trunk and make a
release!

There are also a few semantics I need to check on, but these would be
minor changes.