Add generic type fallback for multiple dispatch!  v1.2.4
When you call spelled_out with an unknown number type, it should tell
you that it's not supported, rather than doing the Pythonic thing of
trying to use it until it fails with a poor error message.  Multiple
dispatch is very convenient to use in this case, but we need to
specify the fallback function manually.  This is what is implemented
in this version of SpelledOut.jl

See merge request #61